Update after PR review -Moved `ReverseGeocodingRequest` out of timefilter package -Added Scaladoc for Geocoding and ReverseGeocoding requests.
