+++ shared/src/main/scala/com/igeolise/traveltimesdk/dto/requests/ReverseGeocodingRequest.scala
@@ -1,15 +1,21 @@
-package com.igeolise.traveltimesdk.dto.requests.timefilter
+package com.igeolise.traveltimesdk.dto.requests
 
 import cats.Monad
 import com.igeolise.traveltimesdk.dto.common.Coords
-import com.igeolise.traveltimesdk.dto.requests.RequestUtils
 import com.igeolise.traveltimesdk.dto.requests.RequestUtils.TravelTimePlatformRequest
-import com.igeolise.traveltimesdk.dto.responses.{GeoJsonResponse, TravelTimeSdkError}
 import com.igeolise.traveltimesdk.dto.responses.common.GeocodingResponseProperties
+import com.igeolise.traveltimesdk.dto.responses.{GeoJsonResponse, TravelTimeSdkError}
 import com.softwaremill.sttp.{sttp, Request, Uri}
 import com.igeolise.traveltimesdk.json.reads.GeocodingReads._
-import scala.language.higherKinds
 
+/**
+  * Attempt to match a latitude, longitude pair to an address.
+  *
+  * @param coordinates Coordinates to match.
+  * @param withinCountry Country code in ISO 3166-1 alpha-2 or alpha-3
+  *                      [[https://en.wikipedia.org/wiki/List_of_ISO_3166_country_codes Country codes]].
+  *                      Useful when performing reverse geocoding requests near a border.
+  */
 case class ReverseGeocodingRequest(
   coordinates: Coords,
   withinCountry: Option[String]
